database: Refactored main database.go file.
